Convert thebasics-analyzer tool into a generic analyzer for any module data
